requirements.txt
google-generativeai
rich
trafilatura
python-dotenv
lxml_html_clean
